requirements.txt
streamlit~=1.43.1
google-generativeai
python-dotenv~=1.0.1
protobuf~=5.29.3